requirements.txt
charset-normalizer
cleanup
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
jsonschema
jsonschema-specifications
Markdown
pillow
psycopg2-binary
PyJWT
